Keep client MatchTimers in sync with the source of truth (the server)
This helps keep the timers accurate in the face of jitter, and also
allows clients that connect during a match to operate correctly.

Note that this is not sufficient for high-latency network environments.

I renumbered the MatchMode enum entries to match their order. While
there is pre-existing code for the FRC FCS that makes assumptions about
the MatchMode order, it actually made bad assumptions about the order
that were incorrect then, but correct now.